Add some member routes. PUT routes added. GET routes... have prompted a lot of thought. The generic GET route potentially returns a lot of data. Actions, History, etc. This requires some additional thought. For instance... actons filter... defines all of hte functionality of the app! So... at this point I don't have that list.
Second... thining through that list of actions... they aren't really importable
from 3rd party systems. It's SO different than what the traditional application
does. It's not dumb. Do I want to let 3rd Party Appls dumb down what we can do,
for the sake of a simple interface? My thought was create a simple app that
interfaced well, so it didn't have to be a be-all-end-all of budget managment.

And maybe it doesn't. But the idea that I can just import dumb data and still
have a useful application--doesn't hold up very well. I *should* know this from
the SPWR experience. Going the traditional route here--will just re-invent the
same sort of problem everyone is alreay struggling with.